requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.51
botocore==1.35.51
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
Django==5.1.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.2
idna==3.10
image==1.5.33
imageio==2.36.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kagglehub==0.3.3
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.24.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyarrow==18.0.0
pyglet==1.5.29
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.6
trimesh==4.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.17.1
